+++ strings/utils.go
@@ -27,19 +27,20 @@ package strings
 
 import "strings"
 
-// ToScreamingDelimited converts a string to SCREAMING.DELIMITED.SNAKE.CASE
-// (in this case `delimiter = '.'; screaming = true`)
-// or delimited.snake.case
-// (in this case `delimiter = '.'; screaming = false`)
+// CustomDelimitedCase converts a string to SCREAMING.DELIMITED.SNAKE.CASE if 'delimiter'
+// equals to '.' and 'screaming' equals to true or to delimited.snake.case if 'delimiter'
+// equals to '.' and screaming equals to false.
 //
-// Credits: https://github.com/iancoleman/strcase
+// 'ignore' variable is what you want the function to ignore and do not change.
+//
+// Credits: https://github.com/iancoleman/strcase/blob/master/snake.go#L66
 //
 // Complexity: O(n)
-func ScreamingDelimited(s string, delimiter uint8, ignore string, screaming bool) string {
-	s = strings.TrimSpace(s)
+func CustomDelimitedCase(input string, delimiter uint8, ignore string, screaming bool) string {
+	input = strings.TrimSpace(input)
 	n := strings.Builder{}
-	n.Grow(len(s) + 2) // nominal 2 bytes of extra space for inserted delimiters
-	for i, v := range []byte(s) {
+	n.Grow(len(input) + 2) // nominal 2 bytes of extra space for inserted delimiters
+	for i, v := range []byte(input) {
 		vIsCap := v >= 'A' && v <= 'Z'
 		vIsLow := v >= 'a' && v <= 'z'
 		if vIsLow && screaming {
@@ -51,18 +52,18 @@ func ScreamingDelimited(s string, delimiter uint8, ignore string, screaming bool
 		}
 
 		// treat acronyms as words, eg for JSONData -> JSON is a whole word
-		if i+1 < len(s) {
-			next := s[i+1]
+		if i+1 < len(input) {
+			next := input[i+1]
 			vIsNum := v >= '0' && v <= '9'
 			nextIsCap := next >= 'A' && next <= 'Z'
 			nextIsLow := next >= 'a' && next <= 'z'
 			nextIsNum := next >= '0' && next <= '9'
 			// add underscore if next letter case type is changed
 			if (vIsCap && (nextIsLow || nextIsNum)) || (vIsLow && (nextIsCap || nextIsNum)) || (vIsNum && (nextIsCap || nextIsLow)) {
-				prevIgnore := ignore != "" && i > 0 && strings.ContainsAny(string(s[i-1]), ignore)
+				prevIgnore := ignore != "" && i > 0 && strings.ContainsAny(string(input[i-1]), ignore)
 				if !prevIgnore {
 					if vIsCap && nextIsLow {
-						if prevIsCap := i > 0 && s[i-1] >= 'A' && s[i-1] <= 'Z'; prevIsCap {
+						if prevIsCap := i > 0 && input[i-1] >= 'A' && input[i-1] <= 'Z'; prevIsCap {
 							n.WriteByte(delimiter)
 						}
 					}